Dodanie klasy do metryk
