+++ Program/PodpisBio/Src/FinalScore/PreciseComparison/DynamicTimeWrapping.cs
@@ -13,12 +13,14 @@ namespace PodpisBio.Src.FinalScore
     {
         const bool REMOVEOUTLYING = true;
         DataPreparation dataPrep = new DataPreparation();
+        Metrics metrics = new Metrics();
         public DynamicTimeWrapping() { }
 
         private float calcSimpleDTW(List<float> ts1, List<float> ts2)
         {
             var n = ts1.Count + 1;
             var m = ts2.Count + 1;
+
             float[][] dtw = new float[n][];
             //jagged arrays są szybsze od [,]
             for (int i = 0; i < n; i++)
@@ -64,45 +66,6 @@ namespace PodpisBio.Src.FinalScore
             return dtw[n - 1][m - 1];
         }
 
-        private float EuclidianDistancePoints(Point p1, Point p2, Derivatives d1, Derivatives d2)
-        {
-            var sum = (p1.X - p2.X) * (p1.X - p2.X);
-            sum += (p1.Y - p2.Y) * (p1.Y - p2.Y);
-
-            sum += (p1.Pressure - p2.Pressure) * (p1.Pressure - p2.Pressure);
-
-            sum += (p1.tiltX - p2.tiltX) * (p1.tiltX - p2.tiltX);
-            sum += (p1.tiltY - p2.tiltY) * (p1.tiltY - p2.tiltY);
-
-            return (float)Math.Sqrt(sum);
-        }
-
-        private float EuclidianDistanceDerivatives(Point p1, Point p2, Derivatives d1, Derivatives d2)
-        {
-            float sum = 0;
-            sum += (d1.Velocity - d2.Velocity) * (d1.Velocity - d2.Velocity);
-            sum += (d1.VelocityX - d2.VelocityX) * (d1.VelocityX - d2.VelocityX);
-            sum += (d1.VelocityY - d2.VelocityY) * (d1.VelocityY - d2.VelocityY);
-
-            sum += (d1.Acc - d2.Acc) * (d1.Acc - d2.Acc);
-            sum += (d1.AccX - d2.AccX) * (d1.AccX - d2.AccX);
-            sum += (d1.AccY - d2.AccY) * (d1.AccY - d2.AccY);
-
-            sum += (d1.PressureChange - d2.PressureChange) * (d1.PressureChange - d2.PressureChange);
-
-            // póki tilty nie działają, nie dodaję ich.
-            return (float)Math.Sqrt(sum);
-        }
-
-        private float EuclidianDistance(Point p1, Point p2, Derivatives d1, Derivatives d2)
-        {
-            float sum = 0;
-
-            sum += EuclidianDistancePoints(p1, p2, d1, d2) * EuclidianDistancePoints(p1, p2, d1, d2);
-            sum += EuclidianDistanceDerivatives(p1, p2, d1, d2) * EuclidianDistanceDerivatives(p1, p2, d1, d2);
-
-            return (float)Math.Sqrt(sum);
-        }
 
         public float calcSimilarity(List<float> sgn1Feature, List<float> sgn2Feature)
         {
@@ -140,7 +103,7 @@ namespace PodpisBio.Src.FinalScore
             //Debug.WriteLine("Śr AccY " + derivatives1.Average(x => x.AccY));
 
 
-            return calcMetricDTW(points1, points2, derivatives1, derivatives2, EuclidianDistance);
+            return calcMetricDTW(points1, points2, derivatives1, derivatives2, metrics.METRIC);
         }
 
         public float calcSimilarity(Signature sgn1, Signature sgn2)
